Major internal redesign that has been on the roadmap for years:
* Add internal evalFuture()

The ultimate goal is to make future stay vanilla futures until just before they are launched. Currently, they are materialized to the backend as soon as they are created, but that will change. This change will make it possible to save futures and, say, reuse them in a future R session, or an R session elsewhere.

This redesign will also simplify the role of future backends and make it easier to write new backends. Eventually, there will be a new FutureBackend class that is separated from the current Future class hierarchy.

This is the first step of many towards these goals, but it's was the major hurdle to get over. The remaining steps should be much easier.

All changes are internal and backward compatible. The updated version has been validated towards the current set of 415 reverse package dependencies on CRAN and Bioconductor. Almost of the very rare, odd unintentional corner cases "out there" are still covered. It has also be validated against the existing test suite, include the 'future.tests' package, and more.